requirements.txt
pyalsaaudio
requests
